- deleting with bug
