+++ src/navigaiton/Tab.js
@@ -39,6 +39,9 @@ const TabNav = (props) => {
     setUserCoordinate,
     setFetchingType,
     setloggedIn,
+    handleLiked,
+    liked,
+    deleteLiked
   } = props;
   return (
     <NavigationContainer>
@@ -79,6 +82,7 @@ const TabNav = (props) => {
               city={city}
               setCity={setCity}
               onLocation={onLocation}
+              handleLiked={handleLiked}
             />
           )}
         </Tab.Screen>
@@ -116,7 +120,6 @@ const TabNav = (props) => {
         </Tab.Screen>
         <Tab.Screen
           name="Liked"
-          component={LikedScreen}
           options={{
             headerShown: false,
             tabBarShowLabel: false,
@@ -135,7 +138,9 @@ const TabNav = (props) => {
               );
             },
           }}
-        />
+        >
+            {(props) => <LikedScreen {...props} liked={liked} deleteLikedItem={deleteLiked} />}
+        </Tab.Screen>
         <Tab.Screen
           name="Profile"
           component={ProfileScreen}
